modified the database creation script, added new methods to UserController and create an error object

+++ src/controllers/UsersController.js
@@ -1,11 +1,18 @@
 const knex = require('../database/knex')
-const { hash } = require('bcryptjs')
+const { hash, compare } = require('bcryptjs')
+const AppError = require('../utils/AppError')
 
 class UsersController{
 
     //create a user
     async create(request, response){
         const { name, email, password } = request.body
+
+        const [checkUserExists] = await knex('users').where({ email })
+
+        if(checkUserExists) {
+            throw new AppError(`user's email already registered`, 409)
+        }
 
         const encryptedPassword = await hash(password, 8)
 
@@ -25,6 +32,12 @@ class UsersController{
     async delete(request, response){
         const { id } = request.query
 
+        const [user] = await knex('users').where({ id })
+
+        if(!user){
+            throw new AppError('User not found', 404)
+        }
+        
         await knex('users').where({ id }).delete()
 
         return response.json({message: `User Id ${id} deleted!`})
@@ -36,13 +49,8 @@ class UsersController{
 
         const [ user ] = await knex('users').where({ id })
 
-
-        //Trabalhar melhor esse erro, estudar o throw new error com express
         if(!user){
-            return response.status(404).json({
-                id:id, 
-                message: "User not found!"
-            })
+           throw new AppError('User not found', 404)
         }
 
         return response.json({
@@ -67,6 +75,54 @@ class UsersController{
         return response.json(usersData)
     }
 
+    async update(request, reponse){
+        const { name, email, password, old_password } = request.body
+        const { id } = request.params
+
+        const [ user ] = await knex('users').where({ id })
+        
+        if(!user){
+            throw new AppError('user not found', 404)
+        }
+
+        //change email
+        if(email){
+            
+            const [ userEmailToUpdate ] = await knex('users').where({ email })
+
+            if(userEmailToUpdate && userEmailToUpdate.id !== user.id){
+                throw new AppError('Email already registered', 409)
+            } else {
+                await knex('users').where({ id }).update({email: email})
+    
+                return reponse.json({
+                    message: 'your email has been changed'
+                })
+            }
+        }
+
+        //change password
+        if(password && !old_password){
+            throw new AppError('please enter your old password', 422)
+        }
+
+        if(password && old_password){
+            const comparePasswords = await compare(old_password, user.password)
+
+            if(!comparePasswords){
+                throw new AppError(`passwords don't match`, 409)
+            }
+
+            const encryptedPassword = await hash(password, 8)
+            await knex('users').where({ id }).update({password: encryptedPassword})
+            return reponse.json({
+                message: 'password has been changed'
+            })
+
+        }
+
+    }    
 }
+
 
 module.exports = UsersController
